Split out client ivc verifier
