+++ barretenberg/cpp/src/barretenberg/stdlib/client_ivc_verifier/CMakeLists.txt
@@ -1,0 +1,8 @@
+barretenberg_module(
+    stdlib_client_ivc_verifier
+    stdlib_circuit_builders
+    goblin
+    translator_vm_recursion
+    eccvm_recursion
+    client_ivc
+)
